fix(index): relocate adjacency tests to unit test module for coverage
Moved `test_import_csr_*` tests from `tests/sentry_adjacency_validation.rs` to `src/index/adjacency.rs` `mod tests`.
This ensures that Codecov correctly detects coverage for the validation logic added to `import_csr`.

Deleted `tests/sentry_adjacency_validation.rs` as it is no longer needed.
